% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRNA96well.R
\name{gRNA96well}
\alias{gRNA96well}
\title{Create gRNA dataframes for 96-well plate ordering}
\usage{
gRNA96well(df, outputpath, wells = 93, nmax = 3)
}
\arguments{
\item{df}{Dataframe output from gRNAmismatchFilter}

\item{outputpath}{Path where you want to save the split dataframes}

\item{wells}{Maximum number of rows each dataframe will have. Default is 93.}

\item{nmax}{Number of gRNAs per gene you want to select. Default is 3.}
}
\value{
Saves dataframes with maximum number of rows specified by wells as .csv
files and excel files for IDT ordering in the outputpath directory
}
\description{
This function loads the output dataframe of gRNAmismatchFilter.
It will split your dataframe into several dataframes for ordering
96-well plates from IDT. e.g., if you specify wells = 93 and you have
215 rows, it will generate 3 dataframes and save them as .csv in the
outputpath directory.
}
