% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRNAatgcFilter.R
\name{gRNAatgcFilter}
\alias{gRNAatgcFilter}
\title{Filter ATGC\% content from gRNA dataframe}
\usage{
gRNAatgcFilter(
  df,
  Arange = 15:35,
  Trange = 15:35,
  Crange = 15:35,
  Grange = 15:35
)
}
\arguments{
\item{df}{Dataframe output from RNA Fold predictions}

\item{Arange}{Percentage range of As for filtering. Default is 15:35.}

\item{Trange}{Percentage range of Ts for filtering. Default is 15:35.}

\item{Crange}{Percentage range of Cs for filtering. Default is 15:35.}

\item{Grange}{Percentage range of Gs for filtering. Default is 15:35.}
}
\value{
Generates a dataframe with filtered gRNAs
}
\description{
This function filters ATGC\% from the output dataframe by
RNA Fold Scores.
}
