% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRNAatgcPlot.R
\name{gRNAatgcPlot}
\alias{gRNAatgcPlot}
\title{Plot ATGC\%}
\usage{
gRNAatgcPlot(df)
}
\arguments{
\item{df}{Dataframe output from RNA Fold predictions}
}
\value{
Creates a ggplot
}
\description{
This function makes a boxplot of ATGC\% content of all gRNAs.
Since this is a ggplot2 object, you can modify the plot by
adding + and ggplot2 code.
}
