% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRNAScorePlot.R
\name{gRNAScorePlot}
\alias{gRNAScorePlot}
\title{Plot RNA Fold scores}
\usage{
gRNAScorePlot(df)
}
\arguments{
\item{df}{Dataframe output from RNA Fold predictions}
}
\value{
Creates a ggplot
}
\description{
This function makes a boxplot of RNA Fold scores of all gRNAs.
Since this is a ggplot2 object, you can modify the plot by
adding + and ggplot2 code.
}
