% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRNAutrperGene.R
\name{gRNAutrperGene}
\alias{gRNAutrperGene}
\title{Plot number of gRNAs in different transcript regions}
\usage{
gRNAutrperGene(df)
}
\arguments{
\item{df}{Dataframe output from RNA Fold predictions}
}
\value{
Creates a ggplot
}
\description{
This function makes a histogram of gRNAs hits per transcript region.
Since this is a ggplot2 object, you can modify the plot by
adding + and ggplot2 code.
}
