% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAFoldOutput.R
\docType{data}
\name{RNAFoldOutput}
\alias{RNAFoldOutput}
\title{Example dataframe}
\format{
A data frame with 20427 rows and 12 columns:
\describe{
Example dataframe output from gRNADesign design module https://github.com/hzuzu/gRNADesign.
}
}
\source{
gRNADesign design module https://github.com/hzuzu/gRNADesign
}
\usage{
RNAFoldOutput
}
\description{
An example dataframe output from gRNADesign design module https://github.com/hzuzu/gRNADesign.
}
\keyword{datasets}
