% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRNAmismatchFilter.R
\name{gRNAmismatchFilter}
\alias{gRNAmismatchFilter}
\title{Filter gRNAs based on mismatches to other transcripts}
\usage{
gRNAmismatchFilter(path, data, extension = "csv")
}
\arguments{
\item{path}{Path to a folder containing all files from gRNADesign module mismatch}

\item{data}{Dataframe of gRNAs selected to go to gRNADesign module mismatch}

\item{extension}{Extension of files in path. Default is csv}
}
\value{
Generates a dataframe with filtered gRNAs
}
\description{
This function loads the output dataframes from gRNADesign module mismatch.
It will filter gRNAs and keep only the ones that have 4 or more
mismatches to any other zebrafish transcript.
}
