% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRNAperGene.R
\name{gRNAperGene}
\alias{gRNAperGene}
\title{Plot number of gRNAs per gene}
\usage{
gRNAperGene(df, binwidth = 5)
}
\arguments{
\item{df}{Dataframe output from RNA Fold predictions}

\item{binwidth}{Specifies binwidth from geom_histogram}
}
\value{
Creates a ggplot
}
\description{
This function makes a histogram of gRNAs hits per gene.
Since this is a ggplot2 object, you can modify the plot by
adding + and ggplot2 code.
}
