% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRNASelection.R
\name{gRNASelection}
\alias{gRNASelection}
\title{Selects gRNAs based on their RNA Fold scores}
\usage{
gRNASelection(df, minN = 10, randomN = 8)
}
\arguments{
\item{df}{Dataframe containing RNA Fold scores}

\item{minN}{Cut-off for minimum number of gRNAs every transcript should have. Default is 10}

\item{randomN}{Number of gRNAs to be selected apart from Top and Bottom RNA Fold scores. Default is 8}
}
\value{
Generates a dataframe with selected gRNAs
}
\description{
This function uses RNA Fold scores to select top, bottom and random
gRNAs. type2 column specifies criteria of choice to be either 'TOP' and 'BOT' scores,
'RANDOM' if randomly selected and 'Less than 10' if less than 10 gRNAs exist for that gene
}
